[java] application refactor
